LED Roulette, a new direction
After a little work it was clear that the effort involved in
going through the blog posts, which are considerably out of date,
was not going to be worth the effort. Embedded rust is moving very
fast. It's better to use the LED roulette application described in
the Discovery book as a learning vehicle